+++ src/index.ts
@@ -11,6 +11,5 @@ export * from "./lib/helpers/dateUtils";
 export * from "./lib/helpers/mui";
 export * from "./lib/helpers/typeahead";
 export * from "./lib/hooks/useDebounceHook";
-export * from "./lib/hooks/useSafeNameId";
 
 export { useFormContext } from "./lib/context/FormContext";