+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Export typeahead helpers and `useDebounceHook` hook.
+
+### Fixed
+
+- Hide placeholder on multiple `AsyncTypeAheadInput` and `StaticTypeAheadInput` when at least one option is selected.
+
 ## [3.1.1] - 2025-04-10
 
 ### Fixed